requirements.txt
numpy==2.2.6
pyyaml==6.0.2
scikit-learn==1.7.2
scipy==1.16.2
tqdm==4.67.1
seaborn==0.13.2
matplotlib==3.10.6
ipykernel==6.30.1   
ipython==9.5.0
spgl1>=0.0.2 # Needed for SRMD
PyWavelets==1.8.0 # Needed for IRCNN+